% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handling_consecutive_regions.R
\name{assign_beta_merged_regions}
\alias{assign_beta_merged_regions}
\title{Assign Beta Values to Merged Regions}
\usage{
assign_beta_merged_regions(regions, cpgs, betavalue)
}
\arguments{
\item{regions}{A data frame containing merged region information, including columns for `chr`, `start`, `end`, and `region_id`.}

\item{cpgs}{A data frame containing CpG data with columns for `chr`, `start`, `end`, and `cpg_id`, which are used to map CpGs to their respective regions.}

\item{betavalue}{A data frame or matrix containing beta values for CpGs. The row names should correspond to the `cpg_id` of the CpGs.}
}
\value{
A list containing two data frames:
\describe{
  \item{regions_cpgs}{A data frame containing merged regions with associated CpG information, including a `cpg_id` and `region_id`.}
  \item{regions_merged_cpg_beta}{A data frame containing the merged regions with their associated beta values. The columns include `region_id`, `cpg_id`, and beta values for each CpG.}
}
}
\description{
This function assigns beta values to merged regions based on CpG locations. It associates each CpG in the provided beta value data with its corresponding region and returns the regions with their associated beta values.
}
\details{
The function first maps CpGs to their respective regions by using the provided `regions` and `cpgs` data. It then merges the CpG data with the beta values based on `cpg_id` and associates each CpG with the appropriate region.

The function outputs two data frames:
\itemize{
  \item The `regions_cpgs` data frame contains the merged regions with CpG information, including a `region_id`.
  \item The `regions_merged_cpg_beta` data frame includes the merged regions with the corresponding beta values for each CpG.
}
}
