% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_differential_methylation.R
\name{combine_into_df}
\alias{combine_into_df}
\title{Combine Results into a Single Data Frame}
\usage{
combine_into_df(DM_results, output_folder)
}
\arguments{
\item{DM_results}{A list containing differential methylation results.}

\item{output_folder}{A character string specifying the directory where the output file should be saved.}
}
\value{
A combined data frame where each row represents a genomic location and each column 
contains diffmeth and adjusted p-value
for different comparisons.
}
\description{
This function merges multiple differential methylation results into a single data frame.
The resulting data frame contains columns for diff.meth and adjusted p-values 
(qvalue) for each comparison.
}
\details{
- The function ensures that genomic locations are properly aligned across all comparisons.
- The resulting data frame is saved as an `.RData` file in the specified output folder.
}
\examples{
\dontrun{
  combine_into_df(DM_results, output_folder = "results/")
}

}
