% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betavalues_mean_variance.R
\name{compute_beta_sd_regions}
\alias{compute_beta_sd_regions}
\title{Compute the Standard Deviation of Beta Values Across Regions and Filter Based on Threshold}
\usage{
compute_beta_sd_regions(regions_cpgs, groups, meth = "betavalues")
}
\arguments{
\item{regions_cpgs}{A data frame or data.table containing methylation data for CpG regions.
The data should include columns for methylation values across different experimental groups, 
and a column `region_id` identifying the CpG region and a `cpg_id` column that map the CpGs inside the region.}

\item{groups}{A character vector of group labels to search within column names of `regions_cpgs` (e.g., control sample identifiers).}

\item{meth}{A character string specifying the type of methylation data. Currently, only "betavalues" is supported.
Default is "betavalues".}
}
\value{
A dataframe containing the `region_id` and the SD of the experimental groups.
}
\description{
This function computes the standard deviation (SD) of Beta values for CpG regions across different groups of samples.
It also visualizes the distribution of SD values across the groups.
}
\details{
The function computes the standard deviation (SD) for Beta values across different groups (replicates) for each CpG region. 
It provides histograms to visualize the SD distributions average across samples.

The function assumes that `regions_cpgs` includes a column `region_id` for CpG region identifiers and columns 
for methylation values, which are grouped by experimental conditions (e.g., different time points or sample types).
}
