% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handling_consecutive_regions.R
\name{compute_cpg_density}
\alias{compute_cpg_density}
\title{Compute CpG Density for Regions}
\usage{
compute_cpg_density(df_list, window_size = 300)
}
\arguments{
\item{df_list}{A data frame or a list of data frames, where each data frame contains CpG information. Each data frame must have columns `chr`, `start`, and `end`, representing the chromosome and the start and end positions of the CpGs.}

\item{window_size}{A numeric value indicating the window size (in base pairs) for computing CpG density. Default is 300.}
}
\value{
A data frame (or a list of data frames) containing the CpG density for each window, with the following columns:
\describe{
  \item{chr}{Chromosome of the CpG region.}
  \item{start}{Start position of the window.}
  \item{end}{End position of the window, calculated as `start + window_size`.}
  \item{num_cpgs}{The number of CpGs within the window.}
}
}
\description{
This function calculates the CpG density for regions within a specified window size. The CpG density is computed for each region by counting the number of CpGs within the defined window.
}
\details{
This function calculates the CpG density within a specified window size for each CpG in a given data frame. The function can either be applied to a single data frame or to a list of data frames. The results are sorted in descending order by the number of CpGs in each window.

The function works by iterating through each CpG in the input data and defining a window around each CpG. It then counts how many CpGs fall within the defined window and stores the results.
}
\examples{
# Example data frame with CpG data
cpg_data <- data.frame(chr = rep(1, 5),
                       start = c(100, 200, 300, 400, 500),
                       end = c(100, 200, 300, 400, 500),
                       stringsAsFactors = FALSE)

# Compute CpG density with a window size of 300
cpg_density <- compute_cpg_density(cpg_data, window_size = 300)

}
