% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handling_consecutive_regions.R
\name{compute_regions_distance}
\alias{compute_regions_distance}
\title{Compute the Distance Between Consecutive Regions}
\usage{
compute_regions_distance(regions_cpgs_consecutive, doplot = TRUE)
}
\arguments{
\item{regions_cpgs_consecutive}{A data.table or data frame containing the CpG regions that are consecutive, with information 
about the region ( `region_id`, `consecutive` flag) and CpG location inside the region `start`, `end`.}

\item{doplot}{A logical value indicating whether a histogram of the computed distances should be plotted. Default is `TRUE`.}
}
\value{
A data.table containing:
\itemize{
  \item `region_pairs`: A string representing the pair of consecutive regions.
  \item `A`: The position of the first CpG in the first region.
  \item `B`: The position of the last CpG in the first region.
  \item `C`: The position of the first CpG in the second region.
  \item `D`: The position of the last CpG in the second region.
  \item `DA`: The distance (in base pairs) between the last CpG of the first region and the first CpG of the second region.
  \item `CB`: The distance (in base pairs) between the first CpG of the second region and the last CpG of the first region.
}
}
\description{
This function computes the distance between two consecutive regions based on their boundary CpG locations.
Specifically, it calculates the distance from the last CpG of the first region to the first CpG of the second region.
Additionally, it checks for any inconsistencies, such as when the end of a region is before its start, and raises a warning.
}
\details{
The function computes the distances between consecutive regions where the `consecutive` flag is `TRUE`. 
It first checks whether the end position of a region appears before its start, raising a warning if so (indicating the 
regions may come from different strands). It then computes the distances between consecutive regions using the boundary 
CpGs of each region (i.e., positions A, B, C, D). A histogram of the distance between consecutive regions (D-A) is 
optionally generated based on the `doplot` argument.
}
