% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_M_values.R
\name{convert_beta_to_Mvalue}
\alias{convert_beta_to_Mvalue}
\title{Convert Beta Values to M-values}
\usage{
convert_beta_to_Mvalue(betavalue, epsilon = 0.01)
}
\arguments{
\item{betavalue}{A \code{data.frame} or matrix containing beta values (typically in the range [0, 100]) for multiple samples.}

\item{epsilon}{A small numeric value added to avoid computing \code{log(0)} or \code{log(Inf)}. Default is \code{1e-2}.}
}
\value{
A \code{data.frame} with the same dimensions as \code{betavalue}, containing the M-value transformation of the input.
}
\description{
This function transforms beta values (proportions in the range [0, 100]) to M-values using a logit2 (base-2 logit)
transformation. It applies a small \code{epsilon} correction to avoid issues with log(0) or division by zero.
}
\details{
The transformation is defined as:
\deqn{M = \log_2 \left(\frac{\beta}{1 - \beta} \right)}
where \eqn{\beta} values are constrained to the range [\code{epsilon}, 1 - \code{epsilon}] and scaled from [0, 100] to [0, 1].
}
\examples{
beta_matrix <- data.frame(
  sample1 = c(0, 50, 100),
  sample2 = c(10, 90, 50)
)
convert_beta_to_Mvalue(beta_matrix)

}
