% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpg_and_gene_annotation.R
\name{download_biomart_annotation}
\alias{download_biomart_annotation}
\title{Download Annotations from BioMart}
\usage{
download_biomart_annotation(
  biomart_database,
  biomart_dataset,
  chromosome_vector,
  selected_attributes,
  attributes_page,
  gene_biotype,
  filter_gene_biotype
)
}
\arguments{
\item{biomart_database}{A string specifying the BioMart database to use (e.g., "ensembl").}

\item{biomart_dataset}{A string specifying the dataset to query (e.g., "hsapiens_gene_ensembl").}

\item{chromosome_vector}{A character vector containing the chromosomes to filter the results.}

\item{selected_attributes}{A character vector of attribute names to be retrieved from BioMart.}

\item{attributes_page}{A character vector specifying the attribute pages to filter relevant attributes.}

\item{gene_biotype}{A string specifying the filter name for gene biotypes (e.g., "biotype").}

\item{filter_gene_biotype}{A character vector containing the gene biotypes to filter for retrieval.}
}
\value{
A named list where each element corresponds to a different attribute page and contains a data frame with the requested annotations.
}
\description{
This function retrieves gene annotations from the BioMart database based on user-specified parameters.
}
\examples{
\dontrun{
biomart_info <- download_biomart_annotation(
  biomart_database = "ensembl",
  biomart_dataset = "hsapiens_gene_ensembl",
  chromosome_vector = c("1", "2", "3"),
  selected_attributes = c("ensembl_gene_id", "hgnc_symbol", "chromosome_name"),
  attributes_page = c("feature_page"),
  gene_biotype = "biotype",
  filter_gene_biotype = c("protein_coding")
)
}

}
