% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handling_consecutive_regions.R
\name{identify_gap_regions}
\alias{identify_gap_regions}
\title{Identify Gap Regions Between Consecutive CpGs}
\usage{
identify_gap_regions(regions_cpgs, doplot = FALSE)
}
\arguments{
\item{regions_cpgs}{A data frame or data.table containing the CpG regions data with columns `start`, `end`, and 
`first_true_after_false`. This data represents the methylation regions along with boundary 
information.}

\item{doplot}{A logical value. If `TRUE`, a histogram will be generated showing the distribution of the gap values 
between consecutive regions. Default is `FALSE`.}
}
\value{
A data.table with an additional column:
\itemize{
  \item `gap`: A numeric column representing the distance (in CpGs) between the end of one region and the start 
              of the next consecutive region. `NA` is assigned for rows that do not mark a gap between consecutive regions.
}
}
\description{
This function identifies the gap between consecutive regions in methylation data. Specifically, it computes the 
distance between the end of one region and the start of the next region that is consecutively attached to it. 
The gap is calculated only for CpGs that mark the boundary between two consecutive regions.
}
\details{
This function calculates the gap between consecutive CpG regions, which is defined as the difference between the 
`start` position of the first CpG in the consecutive region and the `end` position of the last CpG in previous region. It applies the calculation only for CpGs 
where the `first_true_after_false` flag is `TRUE`, indicating that the region marks the beginning of a new consecutive region.
If the `doplot` argument is set to `TRUE`, the function will generate a histogram to visualize the distribution of gap values.
}
