% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handling_consecutive_regions.R
\name{merging_consecutive_regions}
\alias{merging_consecutive_regions}
\title{Merge Consecutive and Non-Consecutive Regions Based on Distance Thresholds}
\usage{
merging_consecutive_regions(filtered_meth_control, tiles, da_thr, cb_thr, ...)
}
\arguments{
\item{filtered_meth_control}{A data frame or data table containing methylation data for control samples, which will be used to investigate adjacent regions.}

\item{tiles}{A data frame or data table containing tile information to map CpGs to regions.}

\item{da_thr}{A numeric value representing the threshold for the `D-A` distance. Regions with a `D-A` distance below this threshold will be merged.}

\item{cb_thr}{A numeric value representing the threshold for the `C-B` distance. Regions with a `C-B` distance below this threshold will be merged.}

\item{...}{Additional arguments that can be passed to the internal functions.}
}
\value{
A data frame containing the combined information for merged regions and non-consecutive regions.
        This includes columns such as chromosome (`chr`), start position (`start`), end position (`end`), and region ID (`region_id`).
        The final dataset contains both merged regions and non-consecutive regions.
}
\description{
This function identifies and merges consecutive regions based on distance thresholds for the `D-A` and `C-B` values. 
It also handles non-consecutive regions by adding them to the final output, providing a combined data set of all regions, both merged and non-merged.
The merging process involves analyzing the distances between regions and applying the specified thresholds to determine which regions should be merged.
}
\details{
The function first investigates the adjacent regions in the data using the `adjacent_regions_investigation` function. 
It then splits the data based on the `D-A` and `C-B` distance thresholds, visualizes the results, and proceeds with the merging process. 
Non-consecutive regions that do not meet the merging criteria are kept in the final dataset.
}
