% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handling_consecutive_regions.R
\name{merging_regions}
\alias{merging_regions}
\title{Merge Regions Based on Distance Thresholds and Optimal Density}
\usage{
merging_regions(pairs_region_summary, regions_cpgs, da_thr, cb_thr)
}
\arguments{
\item{pairs_region_summary}{A data frame containing summaries of region pairs with distance information between consecutive regions.}

\item{regions_cpgs}{A data frame or data table containing CpG region information with corresponding coordinates and region IDs.}

\item{da_thr}{A numeric value representing the threshold for the `D-A` distance. Regions with a `D-A` distance below this threshold will be merged.}

\item{cb_thr}{A numeric value representing the threshold for the `C-B` distance. Regions with a `C-B` distance below this threshold will be merged.}
}
\value{
A data frame containing processed region information, including columns for chromosome (`chr`), start position (`start`), end position (`end`), and region ID (`region_id`). 
        The regions in the returned data frame are either merged or retained as separate based on the specified thresholds.
}
\description{
This function merges regions based on the distances between regions, applying the specified `D-A` and `C-B` distance thresholds. 
It handles different types of merging strategies based on the threshold values and processes regions to either merge them or keep them as separate regions.
}
\details{
The function first splits the region pairs based on the `D-A` and `C-B` distance thresholds using the `split_data_regions_distances` function. 
It then merges regions according to different strategies:
- `D-A` less than the threshold: Regions are pulled together based on the `da_thr` threshold.
- `C-B` less than the threshold: Regions are merged using optimal density considerations.
- `C-B` greater than the threshold: These regions are not merged and are kept as separate regions.
The function returns a combined data frame containing the processed regions.
}
