% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handling_consecutive_regions.R
\name{plot_counts}
\alias{plot_counts}
\title{Plot the Counts of Region Pairs by DA and CB Thresholds}
\usage{
plot_counts(split_data_result, da_thr, cb_thr)
}
\arguments{
\item{split_data_result}{A list containing four data frames, as returned by the `split_data_regions_distances` function. Each data frame represents a subset of region pairs based on the `DA` and `CB` distance thresholds.}

\item{da_thr}{A numeric value representing the threshold for the `D-A` distance.}

\item{cb_thr}{A numeric value representing the threshold for the `C-B` distance.}
}
\value{
A `ggplot` object representing the bar plot.
}
\description{
This function generates a bar plot showing the count of region pairs based on their `D-A` (DA) and `C-B` (CB) distances relative to the specified thresholds. The data is grouped into four conditions based on the comparison of `DA` and `CB` with the given thresholds.
}
\details{
This function is useful for visualizing the distribution of region pairs across the four conditions defined by the `D-A` and `C-B` distance thresholds:
\describe{
  \item{DA <= da_thr & CB <= cb_thr}{Condition where both the `D-A` and `C-B` distances are below or equal to the respective thresholds.}
  \item{DA <= da_thr & CB > cb_thr}{Condition where the `D-A` distance is below or equal to the `da_thr`, but the `C-B` distance is greater than the `cb_thr`.}
  \item{DA > da_thr & CB <= cb_thr}{Condition where the `D-A` distance is greater than the `da_thr`, but the `C-B` distance is below or equal to the `cb_thr`.}
  \item{DA > da_thr & CB > cb_thr}{Condition where both the `D-A` and `C-B` distances are greater than the respective thresholds.}
}

The resulting plot provides a clear visualization of how the data is distributed based on these thresholds.
}
