% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{plot_differential_methylation}
\alias{plot_differential_methylation}
\title{Plot Number of Differentially Methylated CpGs or Regions}
\usage{
plot_differential_methylation(DM_results, filtering, threshold_pval)
}
\arguments{
\item{DM_results}{A list containing differential methylation results from Limma.}

\item{filtering}{A logical value indicating whether to use filtered results (`TRUE`) or unfiltered results (`FALSE`).}

\item{threshold_pval}{A numeric value representing the p-value threshold for filtering (if `filtering = TRUE`).}
}
\value{
A ggplot2 bar plot showing the number of differentially methylated CpGs or regions per sample.
}
\description{
This function generates a bar plot displaying the number of differentially methylated CpGs or regions 
across different samples based on Limma analysis results.
}
\details{
- If `filtering = TRUE`, the plot is based on `DM_results$DM_limma_filtered` with a subtitle indicating the p-value threshold.
- If `filtering = FALSE`, the plot is based on `DM_results$DM_limma` with no subtitle.
- The x-axis represents samples, and the y-axis represents the number of differentially methylated sites.
}
\examples{
\dontrun{
  plot_differential_methylation(DM_results, filtering = TRUE, threshold_pval = 0.05)
}

}
