% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_differential_methylation.R
\name{plot_FC_distribution}
\alias{plot_FC_distribution}
\title{Plot difference in methylation (diff.meth) Distribution from differential methylation}
\usage{
plot_FC_distribution(DM_results, filtering, threshold_pval)
}
\arguments{
\item{DM_results}{A list containing differential methylation results.}

\item{filtering}{A logical value indicating whether to use the filtered results (`TRUE`) 
or unfiltered results (`FALSE`).}

\item{threshold_pval}{A numeric value specifying the p-value threshold used for filtering (if applicable).}
}
\value{
A combined plot displaying histograms of diff.meth values for each comparison.
}
\description{
This function generates histograms of diff.meth from differential 
methylation analysis results, allowing visualization of their distribution.
}
\details{
- If \code{filtering} is `TRUE`, the function plots significant results  
- If \code{filtering} is `FALSE`, the function plots all results.
- The histograms are arranged in a 3-column layout using \code{patchwork}.
}
\examples{
\dontrun{
  plot_FC_distribution(DM_results, filtering = TRUE, threshold_pval = 0.05)
}

}
