% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handling_consecutive_regions.R
\name{split_data_regions_distances}
\alias{split_data_regions_distances}
\title{Split Data Based on Distance Thresholds for D-A and C-B}
\usage{
split_data_regions_distances(df, da_thr, cb_thr)
}
\arguments{
\item{df}{A data frame containing region distance information, including columns for `DA` (D-A distance) and `CB` (C-B distance).}

\item{da_thr}{A numeric value representing the threshold for the `D-A` distance. Data with `DA` less than or greater than this threshold will be split into separate categories.}

\item{cb_thr}{A numeric value representing the threshold for the `C-B` distance. Data with `CB` less than or greater than this threshold will be further split.}
}
\value{
A list containing four data frames:
\describe{
  \item{da_less_cb_less}{Subset of the data where `DA` is less than or equal to `da_thr` and `CB` is less than or equal to `cb_thr`.}
  \item{da_less_cb_more}{Subset of the data where `DA` is less than or equal to `da_thr` and `CB` is greater than `cb_thr`.}
  \item{da_more_cb_less}{Subset of the data where `DA` is greater than `da_thr` and `CB` is less than or equal to `cb_thr`.}
  \item{da_more_cb_more}{Subset of the data where `DA` is greater than `da_thr` and `CB` is greater than `cb_thr`.}
}
}
\description{
This function splits the input data frame into four subsets based on the `D-A` and `C-B` distance thresholds. 
The data is divided into categories where the `D-A` distance is either less than or greater than the `da_thr` threshold, 
and within those, the `C-B` distance is further split based on the `cb_thr` threshold.
}
\details{
This function is useful for categorizing regions based on their `D-A` and `C-B` distances into four distinct groups, which can then be processed separately for further analysis or merging.
}
